fix(PostRepository) change numberOfPosts from Long to Int

+++ app/src/main/java/com/swent/skillswap/model/post/FakePostRepository.kt
@@ -50,7 +50,7 @@ class FakePostRepository : PostRepository {
     }
 
     override suspend fun getMultiplePosts(
-        numberOfPosts: Long,
+        numberOfPosts: Int,
         type: PostType,
         titleContains: String,
         ownerId: String,
