+++ app/src/main/java/com/swent/skillswap/model/post/PostRepository.kt
@@ -39,7 +39,7 @@ interface PostRepository {
      * @throws Exception Database exceptions will be thrown.
      */
     suspend fun getMultiplePosts(
-        numberOfPosts: Long,
+        numberOfPosts: Int,
         type: PostType,
         // optional filters
         titleContains: String = "",
